requirements.txt
streamlit
beautifulsoup4
requests
python-dotenv
email-validator